feat: add settings and update keymaps
This commit does two main things. First, it adds new settings in the
settings.json file for Visual Studio Code. It disables the inline tip for
the 'continue' command and sets the default formatter for JavaScript files
to 'esbenp.prettier-vscode'.

Second, it updates the keymaps in the copilot.lua file for Neovim. It
removes the keymap that was previously set for the 'CopilotChatCommitStaged'
command in insert mode.